Move to ESM only.
