+++ .gitignore
@@ -12,4 +12,4 @@ coverage
 *.old
 @type
 
-/lib/**/*.js
+/lib
