--- ava.config.cjs
@@ -1,14 +1,0 @@
-module.exports = {
-  environmentVariables: {
-    TS_NODE_PROJECT: "tsconfig.ava.json"
-  },
-  extensions: [
-    "ts"
-  ],
-  require: [
-    "ts-node/register",
-  ],
-  files: [
-    "lib/**/*.test.ts"
-  ]
-}
