--- lib/index.ts
@@ -1,3 +1,0 @@
-export * from "./FormData"
-export * from "./Blob"
-export * from "./File"
