+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "formdata-node",
   "version": "4.3.2",
+  "type": "module",
   "description": "Spec-compliant FormData implementation for Node.js",
   "repository": "octet-stream/form-data",
   "sideEffects": false,
@@ -14,28 +15,25 @@
   ],
   "author": "Nick K. <io@octetstream.me>",
   "license": "MIT",
-  "main": "./lib/cjs/index.js",
-  "module": "./lib/esm/browser.js",
-  "browser": "./lib/cjs/browser.js",
+  "main": "./lib/index.js",
+  "module": "./lib/browser.js",
+  "browser": "./lib/browser.js",
   "exports": {
     "./package.json": "./package.json",
     ".": {
       "node": {
         "types": "./@type/index.d.ts",
-        "import": "./lib/esm/index.js",
-        "require": "./lib/cjs/index.js"
+        "import": "./lib/index.js"
       },
       "browser": {
         "types": "./@type/browser.d.ts",
-        "import": "./lib/esm/browser.js",
-        "require": "./lib/cjs/browser.js"
+        "import": "./lib/browser.js"
       },
-      "default": "./lib/esm/index.js"
+      "default": "./lib/index.js"
     },
     "./file-from-path": {
       "types": "./@type/fileFromPath.d.ts",
-      "import": "./lib/esm/fileFromPath.js",
-      "require": "./lib/cjs/fileFromPath.js"
+      "import": "./lib/fileFromPath.js"
     }
   },
   "types": "./@type/index.d.ts",
@@ -55,10 +53,9 @@
     "coverage": "c8 npm test",
     "report:html": "c8 -r=html npm test",
     "ci": "c8 npm test && c8 report --reporter=json",
-    "build:esm": "ttsc --project tsconfig.esm.json",
-    "build:cjs": "ttsc --project tsconfig.cjs.json",
-    "build:types": "ttsc --project tsconfig.d.ts.json",
-    "build": "npm run build:esm && npm run build:cjs && npm run build:types",
+    "build:esm": "tsc",
+    "build:types": "tsc --project tsconfig.d.ts.json",
+    "build": "npm run build:esm && npm run build:types",
     "test": "ava --fail-fast",
     "cleanup": "npx rimraf @type \"lib/**/*.js\"",
     "prepare": "npm run cleanup && npm run build",
@@ -72,7 +69,6 @@
     "@types/sinon": "10.0.11",
     "@typescript-eslint/eslint-plugin": "4.32.0",
     "@typescript-eslint/parser": "4.32.0",
-    "@zoltu/typescript-transformer-append-js-extension": "1.0.1",
     "ava": "4.0.1",
     "c8": "7.11.0",
     "eslint": "7.32.0",
@@ -86,7 +82,6 @@
     "pinst": "3.0.0",
     "sinon": "13.0.1",
     "ts-node": "10.5.0",
-    "ttypescript": "1.5.13",
     "typescript": "4.5.5"
   },
   "dependencies": {
