+++ src/BlobPart.ts
@@ -1,3 +1,3 @@
-import type {Blob, BlobLike} from "./Blob"
+import type {Blob, BlobLike} from "./Blob.js"
 
 export type BlobPart = BlobLike | Blob | Uint8Array
