+++ src/browser.ts
