+++ src/deprecateConstructorEntries.ts
