+++ src/isFunction.ts
