+++ src/isPlainObject.ts
