+++ tsconfig.ava.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.cjs.json",
+  "extends": "./tsconfig.json",
   "ts-node": {
     "preferTsExts": true,
     "transpileOnly": true,
