+++ tsconfig.eslint.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.cjs.json",
+  "extends": "./tsconfig.json",
   "exclude": [
     "node_modules"
   ],
